backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
face-recognition==1.3.0
pillow==10.1.0
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1
pydantic==2.5.0
opencv-python==4.8.1.78
